added menu to game
will now have a main menu
