+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,15 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".Playervsplayer"
+            android:label="@string/title_activity_playervsplayer"
+            android:parentActivityName=".MainActivity"
+            android:theme="@style/AppTheme.NoActionBar">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="com.cs454.connect4.MainActivity" />
+        </activity>
     </application>
 
 </manifest>
